backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg[binary]==3.2.3
requests==2.32.3
cdsapi==0.7.7
google-cloud-storage==2.18.2
xarray==2025.1.2
netCDF4==1.7.2
pandas==2.2.3
pyarrow==18.1.0
